Find test cases recursively in test data (#11)
We did not recursively look for example files in the test data. Instead,
we only looked at the first-level directory.

This patch makes the test scripts look recursively by globbing
``**/*.json or xml``.

As soon as we fixed the search, a couple of serious errors were
revealed. We fixed the issues in aas-core-codegen. Notably, these
changes were relevant:

* [aas-core-codegen 3fea5b26]
* [aas-core-codegen e2a7a806]

Please see the changes in aas-core-codegen if you need more details on
the issues.

[aas-core-codegen 3fea5b26]: https://github.com/aas-core-works/aas-core-codegen/commit/3fea5b26
[aas-core-codegen e2a7a806]: https://github.com/aas-core-works/aas-core-codegen/commit/e2a7a806