Sync naming with the published schemas for V3RC02
Unfortunately, we had a bug in [aas-core-codegen e2a7a806] (2022-10-30)
and prior versions where we generated the names for JSON and RDF
schemas based on a hard-wired list of abbreviations in
*aas-core-codegen* instead of relying on the naming in aas-core-meta.
This list has not been updated before publishing the schemas, so we
have to stick with the unexpected casing in the names. To maintain
the compatibility with the schemas, we allow for capitalization of
certain abbreviations where uppercase would be expected. Concretely,
this was the case with abbreviations "AAS" and "XSD".

In this patch, we rename the class names to match the published schemas
although the names are not pythonic (``AasSubmodelElements`` instead of
``AASSubmodelElements``). While this is ugly, it gives us backwards
compatibility in the future versions of aas-core-meta and
aas-core-codegen when it comes to generating Python code.

This code has been generated using:
* [aas-core-meta d180dbb] and
* [aas-core-codegen 44945c18].

[aas-core-codegen e2a7a806]: https://github.com/aas-core-works/aas-core-codegen/commit/e2a7a806
[aas-core-codegen 44945c18]: https://github.com/aas-core-works/aas-core-codegen/commit/44945c18
[aas-core-meta d180dbb]: https://github.com/aas-core-works/aas-core-meta/commit/d180dbb