Remove TODO from "Getting Started" (#19)
Originally, we wanted to put Python snippets from the docs on an online
REPL site. It turned out that there are no good sites on which we count
for long-term support. As it is fairly easy to copy/paste the snippets
and run them in a Python console, we decided to leave online REPL be.